+++ CHANGELOG.md
@@ -9,6 +9,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 ### Added
 - Alignment plot and event exlusion allows to read or compute cc values
 - `relmt plot-spectra waveform_file` with option to look at different bandpass files
+- Config option 'max_s_amplitude_misfit' to set a different misfit cutoff for
+  S amplitudes
 
 ## [0.3.1-beta] - 2025-12-12